+++ extend.go
@@ -31,6 +31,13 @@ type Extender struct {
 	// Title is the title of the table of contents section.
 	// Defaults to "Table of Contents" if unspecified.
 	Title string
+
+	// MaxDepth is the maximum depth of the table of contents.
+	// Headings with a level greater than the specified depth will be ignored.
+	// See the documentation for MaxDepth for more information.
+	//
+	// Defaults to 0 (no limit) if unspecified.
+	MaxDepth int
 }
 
 // Extend adds support for rendering a table of contents to the provided
@@ -39,7 +46,8 @@ func (e *Extender) Extend(md goldmark.Markdown) {
 	md.Parser().AddOptions(
 		parser.WithASTTransformers(
 			util.Prioritized(&Transformer{
-				Title: e.Title,
+				Title:    e.Title,
+				MaxDepth: e.MaxDepth,
 			}, 100),
 		),
 	)